fix: fix path with a plus character not matched
The root cause for the problem is that when `consumedPath` contains a regexp character, then the regexp `new RegExp('^' + consumedPath, 'i')` doesn't match `segment`.

For example when `consumedPath` is `/foo+bar` then the regexp is `^/foo+bar` and that doesn't match to `"/foo+bar"`. So it is better to avoid the use of regexp here so we don't need to escape user input.